% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice.R
\name{invoice}
\alias{invoice}
\title{Create an invoice object}
\usage{
invoice(month = NULL, year = lubridate::year(lubridate::today()),
  since = if (is.null(month)) {     d <- lubridate::today() - months(1)  
    lubridate::day(d) <- 1     d } else
  lubridate::ymd(sprintf("\%s-\%s-01", year, month)), until = since +
  months(1) - lubridate::days(1),
  client = getOption("invoicer_default_client", "without client"),
  date = until + lubridate::days(1),
  address = getOption("invoicer_default_address", ""), expenses = 0,
  template_key = getOption("invoicer_template_key",
  stop("template key must be specified")), inv_num = 0)
}
\arguments{
\item{month}{specify month for a one-month invoice}

\item{year}{invoice year}

\item{since}{if month is not specified, beginning of invoice period}

\item{until}{end of invoice period}

\item{client}{client name}

\item{date}{invoice date, defaults to the day after until}

\item{address}{client address}

\item{expenses}{amount of expenses for current month}

\item{template_key}{the google drive key for the template to use.
must be in your google drive}

\item{inv_number}{invoice number/code}
}
\value{
an invoice object, a tibble with the following attributes:
\itemize{
   \item {fields} {the various invoice-specific values, such as date
   and address}
   \item {tmpl} {google drive document key of the template}
   \item {url} {url of the google sheet generated}
}
}
\description{
Create an invoice object
}
\examples{
\dontrun{
# create an invoice for the latest complete calendar month using the default
# client, address, and template
invoice()

# create an invoice for this month to date
invoice(
  since = lubridate::floor_date(today(),"month"),
  until = lubridate::today()
)
}

}
