% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_bill_client.R
\name{invoicer_bill_client}
\alias{invoicer_bill_client}
\title{Generate and send any outstanding invoices to the client}
\usage{
invoicer_bill_client(client = "iFixit",
  first_bill_date = as.Date("2018-09-24"), billing_period = 14)
}
\arguments{
\item{client}{client name}

\item{first_bill_date}{date to send first bill}

\item{billing_period}{length of billing period (in days)}
}
\description{
Generate and send any outstanding invoices to the client
}
