% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_billing_schedule.R
\name{invoicer_billing_schedule}
\alias{invoicer_billing_schedule}
\title{Create a billing schedule for a client (through today)}
\usage{
invoicer_billing_schedule(client = "iFixit",
  first_bill_date = as.Date("2018-09-24"), billing_period = 14)
}
\arguments{
\item{client}{client name (character/scalar)}

\item{first_bill_date}{date of the first invoice (Date/scalar)}

\item{billing_period}{length of billing period in days (numeric/scalar)}
}
\description{
Create a billing schedule for a client (through today)
}
