% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_check_data.R
\name{invoicer_check_data}
\alias{invoicer_check_data}
\title{Check if an invoice should be created based on the data provided}
\usage{
invoicer_check_data(x)
}
\arguments{
\item{x}{invoice data}
}
\description{
Check if an invoice should be created based on the data provided
}
