% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_create.R
\name{invoicer_create}
\alias{invoicer_create}
\title{Generate an HTML invoice document}
\usage{
invoicer_create(rds_file = NULL, date_due = Sys.Date() + 14,
  include_dates = TRUE, aggregate = FALSE)
}
\arguments{
\item{rds_file}{full path to .rds file containing invoice data}

\item{date_due}{due date for invoice (Date/scalar)}

\item{include_dates}{include a 'Date' column in invoice?}

\item{aggregate}{aggregate projects over multiple days}
}
\description{
Generate an HTML invoice document
}
