% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_email.R
\name{invoicer_email}
\alias{invoicer_email}
\title{Send an invoice via email}
\usage{
invoicer_email(x, id)
}
\arguments{
\item{x}{invoice object}

\item{id}{invoice id}
}
\description{
Send an invoice via email
}
