% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_email.R
\name{invoicer_email_send}
\alias{invoicer_email_send}
\title{Compose/send an invoice email}
\usage{
invoicer_email_send(x, preview = FALSE)
}
\arguments{
\item{x}{invoice data}

\item{preview}{preview email}
}
\description{
Compose/send an invoice email
}
