% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_filter_data.R
\name{invoicer_filter_data}
\alias{invoicer_filter_data}
\title{Filter invoice data}
\usage{
invoicer_filter_data(x, client, start_date, end_date)
}
\arguments{
\item{x}{invoice data (\code{\link{invoicer_get_data}})}

\item{client}{client name (character/scalar)}

\item{start_date}{first day of work (Date/scalar)}

\item{end_date}{last day of work (Date/scalar)}
}
\description{
Filter invoice data
}
