% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_formats.R
\name{invoicer_format_link}
\alias{invoicer_format_link}
\title{Convert a github link to a href html tag}
\usage{
invoicer_format_link(url = "https://github.com/iFixit/Stats/issues/280")
}
\arguments{
\item{x}{url}
}
\description{
Convert a github link to a href html tag
}
