% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_formats.R
\name{invoicer_format_links}
\alias{invoicer_format_links}
\title{Convert a comma-separated string of github links to href tags}
\usage{

  invoicer_format_links(urls = "https://github.com/iFixit/Stats/issues/280, https://github.com/iFixit/Stats/issues/290")
}
\arguments{
\item{urls}{url(s)}
}
\description{
Convert a comma-separated string of github links to href tags
}
