% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_formats.R
\name{invoicer_format_project_title}
\alias{invoicer_format_project_title}
\title{Create an invoice project title; if a github link is provided, convert to its 'title'}
\usage{
invoicer_format_project_title(title = "This Project")
}
\arguments{
\item{title}{project title}
}
\description{
Create an invoice project title; if a github link is provided, convert to its 'title'
}
