% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_get_data.R
\name{invoicer_get_company}
\alias{invoicer_get_company}
\title{Load company data from googlesheets}
\usage{
invoicer_get_company(key = Sys.getenv("INVOICER_GS_KEY"), gs = NULL)
}
\arguments{
\item{key}{googlesheets key}

\item{gs}{registered googlesheet}
}
\description{
Load company data from googlesheets
}
\note{
requires authentication via googlesheets
}
