% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_get_data.R
\name{invoicer_get_data}
\alias{invoicer_get_data}
\title{Load invoice data from googlesheets}
\usage{
invoicer_get_data(key = Sys.getenv("INVOICER_GS_KEY"))
}
\arguments{
\item{key}{googlesheets key}
}
\description{
Load invoice data from googlesheets
}
\note{
requires authentication via googlesheets
}
