% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_html.R
\name{invoicer_html}
\alias{invoicer_html}
\title{Generate an invoice.html based on the invoice data in 'x'}
\usage{
invoicer_html(x, include_dates = FALSE, aggregate = FALSE)
}
\arguments{
\item{x}{invoicer object}

\item{include_dates}{include a 'Date' column in invoice?}

\item{aggregate}{aggregate projects over multiple days}
}
\value{
html
}
\description{
Generate an invoice.html based on the invoice data in 'x'
}
