% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_html_details.R
\name{invoicer_html_details}
\alias{invoicer_html_details}
\title{Create invoice details (project, hours, price)}
\usage{
invoicer_html_details(x, include_dates = FALSE, aggregate = FALSE)
}
\arguments{
\item{x}{invoicer object}

\item{include_dates}{include a 'Date' column in invoice?}

\item{aggregate}{aggregate projects over multiple days}
}
\value{
html
}
\description{
Create invoice details (project, hours, price)
}
