% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_html_header.R
\name{invoicer_html_header}
\alias{invoicer_html_header}
\title{Create invoice header}
\usage{
invoicer_html_header(x, include_dates = TRUE)
}
\arguments{
\item{x}{invoicer object}

\item{include_dates}{include a 'Date' column in invoice?}
}
\value{
html
}
\description{
Create invoice header
}
