% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_html_info.R
\name{invoicer_html_info}
\alias{invoicer_html_info}
\title{Create invoice contractor/client info}
\usage{
invoicer_html_info(x, include_dates = TRUE)
}
\arguments{
\item{x}{invoicer object}

\item{include_dates}{include a 'Date' column in invoice?}
}
\value{
html
}
\description{
Create invoice contractor/client info
}
