% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_record.R
\name{invoicer_record}
\alias{invoicer_record}
\title{Record data in a google sheet}
\usage{
invoicer_record(key = "1auWdQK9f0db09rVItdgAqMEUHrgYt8x-vt4coQi0g3w",
  gs = NULL, ws, x)
}
\arguments{
\item{key}{googlesheets key}

\item{gs}{registered googlesheet}

\item{ws}{worksheet}

\item{x}{data.frame to insert}
}
\description{
Record data in a google sheet
}
