% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_setup.R
\name{invoicer_setup}
\alias{invoicer_setup}
\title{Run this function to create the googlesheets template}
\usage{
invoicer_setup(name = "AVP Consulting", email = "apileggi20@gmail.com")
}
\arguments{
\item{name}{Company name}
}
\description{
Run this function to create the googlesheets template
}
