% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoicer_summary.R
\name{invoicer_summary}
\alias{invoicer_summary}
\title{Invoice summary}
\usage{
invoicer_summary(x)
}
\arguments{
\item{x}{invoice data from \code{\link{invoicer_filter_data}}}
}
\description{
Invoice summary
}
