% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{wrap a function so that it keeps executing until none of the specified 
errors is returned, probably most useful for functions that try to use the
network}
\usage{
retry(.f, errors = getOption("retry_errors",
  "Error in the HTTP2 framing layer"), max_tries = 10)
}
\arguments{
\item{.f}{function to be wrapped}

\item{errors}{vector of error messages on whose return we should retry}

\item{max_tries}{maximum number of retries}
}
\description{
wrap a function so that it keeps executing until none of the specified 
errors is returned, probably most useful for functions that try to use the
network
}
