% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{invoicer-package}
\alias{invoicer}
\alias{invoicer-package}
\title{invoicer: KopfLab Internal Package for Invoicing}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})

}
\keyword{internal}
